requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pymysql==1.1.1
cryptography==43.0.3
google-generativeai==0.8.3
scikit-learn==1.5.2
numpy==2.1.3
python-dotenv==1.0.1
pydantic==2.9.2
httpx==0.27.2
pydantic-settings==2.6.1
